Make settings-pane a component, move the api call to actions
