+++ packages/devtools-launchpad/src/actions/types.js
@@ -1,0 +1,16 @@
+/**
+ * Flow types
+ * @module actions/types
+ */
+
+/**
+  * Argument parameters via Thunk middleware for {@link https://github.com/gaearon/redux-thunk|Redux Thunk}
+  *
+  * @memberof actions/breakpoints
+  * @static
+  * @typedef {Object} ThunkArgs
+  */
+export type ThunkArgs = {
+  dispatch: () => Promise<any>,
+  getState: () => any
+};
